Resolved conflicts between the incentive-for-low-bond-proposals branch
and feat/protocol/shasta, including:

- Updated IInbox interface to include coreStateHash field in Proposal struct
- Refactored Inbox implementation with new validation and processing logic
- Updated MainnetInbox to use InboxOptimized instead of InboxWithClaimAggregation
- Converted bond amounts from uint256 to uint48/uint96 for gas optimization
- Updated BondManager and ShastaAnchor contracts to use new bond types
- Fixed compilation issues in dependent contracts and tests
- Removed LibDecoder dependency and updated to use LibCodec
- Updated storage layout documentation

+++ packages/protocol/contracts/layer2/based/IBondManager.sol
@@ -66,7 +66,7 @@ interface IBondManager {
     /// @notice Credits a bond to an address
     /// @param _address The address to credit the bond to
     /// @param _bond The amount of bond to credit
-    function creditBond(address _address, uint256 _bond) external;
+    function creditBond(address _address, uint96 _bond) external;
 
     /// @notice Gets the bond balance of an address
     /// @param _address The address to get the bond balance for
